chore(opentelemetry source): split config from log service
